requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
rdkit-pypi==2023.9.5
scikit-learn==1.3.2
numpy==1.26.3
pandas==2.1.4
gunicorn==21.2.0
whitenoise==6.6.0
python-dotenv==1.0.0